### 2019-08-21 #1671 @whedon
[PRE REVIEW]: qnm: A Python package for calculating Kerr quasinormal modes, separation constants, and spherical-spheroidal mixing coefficients
**Submitting author:** @duetosymmetry (<a href="http://orcid.org/0000-0001-7559-9597">Leo C. Stein</a>)
**Repository:** <a href="https://github.com/duetosymmetry/qnm" target ="_blank">https://github.com/duetosymmetry/qnm</a>
**Version:** 0.3.0
