### 2019-08-22 #1671 @xuanxu
:wave: @duncanmmacleod: would you be willing to review this paper? I think your knowledge of Python and gravitational waves makes you a great fit for this submission.
