### 2019-08-27 #1671 @duncanmmacleod
> 👋 @duncanmmacleod: would you be willing to review this paper? I think your knowledge of Python and gravitational waves makes you a great fit for this submission.

Yes, I am happy to review, but will not able to read or contribute anything for at least two (2) weeks due to existing commitments. If you're happy to wait, I'm happy to get to it after September 9.
