### 2019-08-27 eb0c361
Under review at JOSS, update paper Makefile consts. Use microtype.
