+++ docs/paper/Makefile
@@ -8,18 +8,18 @@ LATEXFILE=paper.tex
 # some variables for the paper
 REPOSITORY=https://github.com/duetosymmetry/qnm
 ARCHIVE=https://doi.org/10.5281/zenodo.2593978
-REVIEW=https://github.com/openjournals/joss-reviews/issues/XXX
+REVIEW=https://github.com/openjournals/joss-reviews/issues/1683
 ARCHIVEDOI=https://doi.org/10.5281/zenodo.2593978
-FORMATTEDDOI=10.21105/joss.XXXXX
+FORMATTEDDOI=10.21105/joss.01683
 YEAR=2019
 TITLE="qnm: A Python package for calculating Kerr quasinormal modes, separation constants, and spherical-spheroidal mixing coefficients"
 AUTHOR="Stein"
-ISSUE=XX
-VOLUME=X
-PAGE=XXX
+ISSUE=40
+VOLUME=4
+PAGE=1683
 JOURNALNAME="Journal of Open Source Software"
 PUBLISHED="XXX XXX 2019"
-SUBMITTED="XXX XXX 2019"
+SUBMITTED="21 August 2019"
 PAPERURL=https://www.theoj.org/joss-papers/joss.XXXXX/10.21105.joss.XXXXX.pdf
 
 latex.template:
@@ -32,7 +32,7 @@ joss-logo.png:
 
 latex: $(MDFILE) $(BIBFILE) latex.template joss-logo.png
 	# use pandoc and pandoc-citeproc to create LaTeX source
-	pandoc --filter pandoc-citeproc --bibliography=$(BIBFILE) -V repository="$(REPOSITORY)" -V journal_name=$(JOURNALNAME) -V paper_url="$(PAPERURL)" -V archive_doi="$(ARCHIVEDOI)" -V review_issue_url="$(REVIEW)" -V formatted_doi="$(FORMATTEDDOI)" -V logo_path="joss-logo.png" -V year="$(YEAR)" -V paper_title=$(TITLE) -V citation_author=$(AUTHOR) -V issue="$(ISSUE)" -V volume="$(VOLUME)" -V page="$(PAGE)" -V submitted=$(SUBMITTED) -V published=$(PUBLISHED) -V graphics="true" -V papersize=a4paper -V geometry:margin=1in -t latex -s $(MDFILE) -o $(LATEXFILE) --template latex.template
+	pandoc --filter pandoc-citeproc --bibliography=$(BIBFILE) -V repository="$(REPOSITORY)" -V journal_name=$(JOURNALNAME) -V paper_url="$(PAPERURL)" -V archive_doi="$(ARCHIVEDOI)" -V review_issue_url="$(REVIEW)" -V formatted_doi="$(FORMATTEDDOI)" -V logo_path="joss-logo.png" -V year="$(YEAR)" -V footnote_paper_title=$(TITLE) -V citation_author=$(AUTHOR) -V issue="$(ISSUE)" -V volume="$(VOLUME)" -V page="$(PAGE)" -V submitted=$(SUBMITTED) -V published=$(PUBLISHED) -V graphics="true" -V papersize=a4paper -V geometry:margin=1in -t latex -s $(MDFILE) -o $(LATEXFILE) --template latex.template
 
 pdf: latex
 	pdflatex paper.tex
