### 2019-08-27 5b98e47
Include author info in PDF metadata
