### 2019-08-29 4652fd9
Add bibtex key (arXiv-only until JOSS review finishes)
