### 2019-08-29 323c267
Need a newer version of recommonmark for img alt's, thanks @mcnees!
