+++ docs/_autosummary/qnm.angular.rst
@@ -8,14 +8,24 @@ qnm.angular
    .. rubric:: Functions
 
    .. autosummary::
-
+   
       C_and_sep_const_closest
-      ells
-      give_M_matrix_elem_ufunc
-      l_min
       M_matrix
       M_matrix_elem
       M_matrix_old
+      ells
+      give_M_matrix_elem_ufunc
+      l_min
       sep_const_closest
       sep_consts
       swsphericalh_A
+   
+   
+
+   
+   
+   
+
+   
+   
+   
