+++ docs/_autosummary/qnm.radial.rst
@@ -9,8 +9,18 @@ qnm.radial
 
    .. autosummary::
    
-      sing_pt_char_exps
       D_coeffs
       leaver_cf_inv_lentz
       leaver_cf_inv_lentz_old
       leaver_cf_trunc_inversion
+      sing_pt_char_exps
+   
+   
+
+   
+   
+   
+
+   
+   
+   
