+++ docs/requirements.txt
@@ -1,3 +1,3 @@
 numpydoc
-recommonmark==0.5.0
+recommonmark==0.6.0
 sphinx==1.8.4
