### 2019-08-30 21b0e0e
Simplify examples in README
