### 2019-09-02 #1683 @mattpitkin
Hi @duetosymmetry, one of the requirements of the review to check that the documentation contains

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Could you add a sentence or two stating ow users can contribute/leave feedback, e.g., via Issue or PRs on the package's git repo?
