### 2019-09-02 #1683 @mattpitkin
@duetosymmetry in the bibtex file can you add a `journal` entry to the Isi et al, paper, so that the arXiv number shows up, e.g.:

```
journal = "arXiv:\href{https://arxiv.org/abs/1905.00869}{1905.00869}"
```

(or course this can be updated if the actual journal article comes out in the next couple of weeks)
