### 2019-09-02 #1683 @mattpitkin
@duetosymmetry do you have any plans to add any unit testing, and or a CI builder for the package? These aren't requirements for acceptance in JOSS, but I just wanted to check. I've can run the example in the documentation, but is there any example case for which values that the code can produce that can be checked against any independent published results?
