### 2019-09-02 ccb51bc
Add fake journal to Isi bib entry, for JOSS LaTeX template
