+++ docs/paper/paper.bib
@@ -144,6 +144,7 @@
       author         = "Isi, Maximiliano and Giesler, Matthew and Farr, Will M.
                         and Scheel, Mark A. and Teukolsky, Saul A.",
       title          = "{Testing the no-hair theorem with GW150914}",
+      journal        = "arXiv:\href{https://arxiv.org/abs/1905.00869}{1905.00869}",
       year           = "2019",
       eprint         = "1905.00869",
       archivePrefix  = "arXiv",
