### 2019-09-02 #1683 @duetosymmetry
Thanks for the feedback, @mattpitkin .

1. Added short contribution guidelines in duetosymmetry/qnm@14e7e7a
2. I didn't see the missing arXiv number when I was making with the JOSS template locally -- there must be some subtle difference between the template I fetched and the one that whedon is using. Fixed in duetosymmetry/qnm@ccb51bc
3. I would like to add CI testing, e.g. against Emanuele Berti's precomputed tables, but I'm not sure how best to do this. To make this happen would require writing a test that would download his tarballs, which is ~100Mb, currently hosted on a JHU server. If an installation of `qnm` uses `qnm.download_data()` to speed up the testing, then that would download another ~100Mb, from my web server. Is there a better approach? Maybe pre-selecting a handful of combinations of (s,l,m,n,a), and storing those statically in a test? Is there a python package with a simple example of testing infrastructure that you might point me to? For now, there are only doctests that can be performed by going into docs and performing `make doctest`.
