### 2019-09-03 #1683 @mattpitkin
> 1. Added short contribution guidelines in duetosymmetry/qnm@14e7e7a

Excellent, thanks. I'll tick off the appropriate review check box.
