### 2019-09-03 #1683 @mattpitkin
> 2. I didn't see the missing arXiv number when I was making with the JOSS template locally -- there must be some subtle difference between the template I fetched and the one that whedon is using. Fixed in duetosymmetry/qnm@ccb51bc

Thanks, I can see the arXiv number in the generated pdf now.
