### 2019-09-03 #1683 @mattpitkin
> 3. I would like to add CI testing, e.g. against Emanuele Berti's precomputed tables, but I'm not sure how best to do this. To make this happen would require writing a test that would download his tarballs, which is ~100Mb, currently hosted on a JHU server. If an installation of qnm uses qnm.download_data() to speed up the testing, then that would download another ~100Mb, from my web server. Is there a better approach? Maybe pre-selecting a handful of combinations of (s,l,m,n,a), and storing those statically in a test? Is there a python package with a simple example of testing infrastructure that you might point me to? For now, there are only doctests that can be performed by going into docs and performing make doctest.

It's now pretty simple to integrate CI testing in [Travis](https://travis-ci.org/) for github projects. It's not a huge issue for these tests to download 100Mb data sets each time - often you can get then to CI to download a Docker image for the testing each time, which might be larger. An example of the CI configuration for one of my own package's can be found [here](https://github.com/mattpitkin/psrqpy/blob/master/.travis.yml). This uses [pytest](https://docs.pytest.org/en/latest/) to run a set of test scripts found in the [`test` directory](https://github.com/mattpitkin/psrqpy/tree/master/test).

At it's simplest you could add a `.travis.yml` file to the repo with something like:

```yml
language: python
sudo: required
dist: xenial
branches:
  only:
    master
python:  # run test on both Python 3.5 and 3.6
  - "3.6"
  - "3.7"
install:  # install your package and requirements
  - pip install --upgrade pip
  - pip install -r requirements.txt
  # install packages for testing
  - pip install pytest
  # build the qnm package
  - pip install -e .
script:
  # run the test script
  - pytest
```

Then you could create a `test` directory (which will be automatically looked for by `pytest`) and add a file, e.g., `test_qnm.py` containing something like:

```python
import pytest
import qnm
import numpy as np

class test_qnm(object):
    """
    A class for testing aspects of the qnm module.
    """
    @classmethod
    def setup_class(cls):
        """
        Download the data when setting up the test class.
        """
        qnm.download_data()

    def test_example(self):
        """
        An example of a test
        """
        grav_220 = qnm.modes_cache(s=-2,l=2,m=2,n=0)
        omega, A, C = grav_220(a=0.68)
        assert omega == (0.5239751042900845 - 1j * 0.08151262363119974j)

    def test_example2(self):
        from qnm.radial import leaver_cf_inv_lentz_old, leaver_cf_inv_lentz
        old = leaver_cf_inv_lentz_old(omega=.4 - 0.2j, a=0.02, s=-2, m=2, A=4.+0.j, n_inv=0)
        new = leaver_cf_inv_lentz(omega=.4 - 0.2j, a=0.02, s=-2, m=2, A=4.+0.j, n_inv=0)
        assert np.all([old[i] == new[i] for i in range(3)])
```

Then, if you sign-up for a Travis CI account, you can activate the CI for the qnm git repo and the CI should be run each time you commit a change. Then you can add a [badge](https://docs.travis-ci.com/user/status-images/) to your repo :smile: :name_badge:
