### 2019-09-03 06db0a7
Docs need extra files for contributing guidelines
