### 2019-09-04 ddae3e2
Scaffolding for test infrastructure
