+++ .travis.yml
@@ -1,0 +1,20 @@
+language: python
+sudo: required
+dist: xenial
+branches:
+  only:
+    add-test-infra #master
+python:
+  - "2.7"
+  - "3.6"
+  - "3.7"
+install:  # install your package and requirements
+  - pip install --upgrade pip
+  - pip install -r requirements.txt
+  # install packages for testing
+  - pip install pytest
+  # build the qnm package
+  - pip install -e .
+script:
+  # run the test script
+  - pytest
