### 2019-09-04 #1683 @duetosymmetry
Thanks for the quick tutorial, @mattpitkin . I just integrated with travis successfully! I'll get to writing real tests, but probably not until tomorrow.
