### 2019-09-04 a10eba9
Enable pip caching on travis
