### 2019-09-12 #1683 @duetosymmetry
Update for @mattpitkin: I started building some comparison code between Berti's tables and my code. Unfortunately some of the differences |Berti - me| are greater than the error estimates that my package is reporting. So right now I'm not sure I can trust Emanuele's numbers for these automated tests.
