### 2019-09-15 #1683 @duetosymmetry
Update again: I was able to perform an offline comparison again Greg Cook's tabulated data, and the comparison is very good. Cook's data is on Zenodo, but I'm not going to make an online automated test based on this data, because his dataset is 1 gig. I will put automated testing of the data quality on the back burner for now until I come up with a good solution. I can build some regression tests, which should not be very hard, but this is not a very high priority for me at the moment.
