### 2019-09-15 7853cb7
Bit of documentation cleanup
