### 2019-09-15 cf672b6
Break out decompressing data, add func for clearing disk cache, test
