### 2019-09-16 duetosymmetry/qnm#21 @duncanmmacleod
Data caching is not platform independent
The [`download_data`](https://github.com/duetosymmetry/qnm/blob/v0.3.0/qnm/cached.py#L401-L435) function relies upon `'/'` as the path separator, meaning it is (probably) incompatible with Windows systems. It would be better to use [`pathlib`](https://docs.python.org/3/library/pathlib.html) to generate the paths programmatically:

```python
from pathlib import Path

def download_data(...):
    ...
    base_dir = Path(__file__).parent
    dest = base_dir / filename
    ...

    data_dir = base_dir / "data"
    pickle_files = data_dir.glob("*.pickle")
    ...
```
