### 2019-09-16 #1683 @duncanmmacleod
@duetosymmetry, apologies for the delay in reviewing this work.

Overall this looks very nice, I am particularly pleased with the minimal dependencies and the used of JIT compilation to cache outputs. I have posted the following issues to the target repo:

- https://github.com/duetosymmetry/qnm/issues/20
- https://github.com/duetosymmetry/qnm/issues/21

Once these are addressed (or plans drawn up to address them in the future), I will be happy to approve this.
