### 2019-09-16 #1683 @mattpitkin
@duetosymmetry thanks for your updates. If you could provide a script, and just a short description in the docs, of the offline comparison to a couple of cases in Greg Cook's data table that would satisfy me.
