### 2019-09-17 c9c9c39
Bugfix: argument to optimize.root was treated as positional.
self.tol was supposed to set `tol=self.tol`, but it was being
interpreted as positional, and that position is `args`.  This was
sending self.tol as an extra argument to `__call__` instead of setting
the tolerance!!
