### 2019-09-17 88a354d
Turn on travis on master, make pytest more verbose
