### 2019-09-17 c27ae87
Add Travis badge to README
