### 2019-09-20 #1683 @duetosymmetry
Hi folks, I've created a coLaboratory notebook that fetches Greg Cook's data from Zenodo, loads the `qnm` package, does comparisons between the two, and plots the differences in frequencies and separation constants. You can check it out [here](https://colab.research.google.com/drive/1qIMC1gGzNP_ebg5SOBCHGx7P5jImtXXe#forceEdit=true&sandboxMode=true).
Any suggestions for sharing this notebook? I don't consider it as part of the package. I could make a vanilla jupyter notebook and stick that in some other repo. Or just leave this on coLab, then others don't have to install anything to verify the claims of the package.
