### 2019-09-20 #1683 @duetosymmetry
PS Have already started working on using `pathlib` (today I learned!) and the cache location, probably won't get a chance to fix everything up until this weekend. Too busy with teaching at the moment.
