### 2019-09-20 #1683 @mattpitkin
> Any suggestions for sharing this notebook? I don't consider it as part of the package. I could make a vanilla jupyter notebook and stick that in some other repo. Or just leave this on coLab, then others don't have to install anything to verify the claims of the package.

I'd suggest just adding it to the `qnm` repo in the `notebooks` directory, but if you'd prefer not to then putting it in its own standalone repo (and including with that a link to the CoLab notebook) is probably the best - hopefully a gitlab repo would be a bit more "permanent" that a CoLab notebook.

Whatever you chose, I'd suggest adding a link to, and short description of, the notebook in the `README.md` and/or docs of `qnm`.
