### 2019-09-22 5abc67d
Be explicit about using scipy.optimize.root method hybr, add TODO
