### 2019-09-22 6fed59e
Break out get_cachedir() func, starting to address #20 and #21
