### 2019-09-22 5f99fd8
Use pathlib.Path everywhere possible, closes #21.
