+++ qnm/cached.py
@@ -15,7 +15,6 @@ from __future__ import division, print_function, absolute_import
 
 import logging
 import pickle
-import os
 from pathlib import Path
 try:
     from urllib.request import urlretrieve # py 3
@@ -38,13 +37,11 @@ def get_cachedir():
     """
     Return the location of the cache directory.
 
-    TODO Use pathlib
-
     Returns
     -------
-    string
+    pathlib.Path object
     """
-    return os.path.dirname(os.path.realpath(__file__))
+    return Path(__file__).parent.resolve()
 
 def mode_pickle_path(s, l, m, n):
     """Construct the path to a pickle file for the mode (s, l, m, n)
@@ -65,7 +62,7 @@ def mode_pickle_path(s, l, m, n):
 
     Returns
     -------
-    string
+    pathlib.Path object
       `<cachedir>/data/s<s>_l<l>_m<m>_n<n>.pickle`
 
      """
@@ -82,8 +79,8 @@ def mode_pickle_path(s, l, m, n):
         s_sign, np.abs(s), l,
         m_sign, np.abs(m), n)
 
-    return os.path.abspath(
-        '{}/data/{}'.format(get_cachedir(), filename))
+    pickle_path = get_cachedir() / 'data' / filename
+    return pickle_path.resolve()
 
 def write_mode(spin_seq, pickle_path=None):
     """Write an instance of KerrSpinSeq to disk.
@@ -93,7 +90,7 @@ def write_mode(spin_seq, pickle_path=None):
     spin_seq: KerrSpinSeq
       The mode to write to disk.
 
-    pickle_path: string, optional [default: None]
+    pickle_path: string or pathlib.Path, optional [default: None]
       Path to file to write. If None, get the path from
       :meth:`mode_pickle_path()`.
 
@@ -107,15 +104,17 @@ def write_mode(spin_seq, pickle_path=None):
         pickle_path = mode_pickle_path(spin_seq.s, spin_seq.l,
                                        spin_seq.m, spin_seq.n)
 
-    _the_dir = os.path.dirname(pickle_path)
-    if not os.path.exists(_the_dir):
+    # Convert pickle_path to Path if it's a string
+    pickle_path = Path(pickle_path)
+    the_dir = pickle_path.parent
+    if not the_dir.exists():
         try:
-            os.makedirs(_the_dir)
+            the_dir.mkdir(parents=True, exist_ok=True)
         except:
-            logging.error("Could not create dir {} to store Kerr QNM sequence".format(_the_dir))
+            logging.error("Could not create dir {} to store Kerr QNM sequence".format(the_dir))
 
     try:
-        with open(pickle_path, 'wb') as handle:
+        with pickle_path.open('wb') as handle:
             logging.info("Writing Kerr QNM sequence to file {}".format(pickle_path))
             pickle.dump(spin_seq, handle, protocol=pickle.HIGHEST_PROTOCOL)
     except:
@@ -154,17 +153,17 @@ def load_cached_mode(s, l, m, n):
     pickle_path = mode_pickle_path(s, l, m, n)
 
     try:
-        with open(pickle_path, 'rb') as handle:
+        with pickle_path.open('rb') as handle:
             logging.info("Loading Kerr QNM sequence from file {}".format(pickle_path))
             spin_seq = pickle.load(handle)
     except UnicodeDecodeError as e:
-        with open(pickle_path, 'rb') as handle:
+        with pickle_path.open('rb') as handle:
             logging.info("Loading Kerr QNM sequence from file {}".format(pickle_path))
             spin_seq = pickle.load(handle, encoding='latin1')
     except:
+        logging.warn("Could not load Kerr QNM sequence from file {}.".format(pickle_path))
         if not hasattr(load_cached_mode, "have_warned"):
[…]
