### 2019-09-22 9cfeee1
urlretrieve relies on string instead of Path
