### 2019-09-22 c36cc76
Legacy tarfile also only accepts str instead of Path
