### 2019-09-22 95aba79
Use test for existence of pickle rather than relying on failure
