### 2019-09-22 1c45873
Test bug, glob in wrong dir
