### 2019-09-22 14ed15a
Remove extraneous Path resolve()
