### 2019-09-23 fbae5fe
Separate tolerances for cont frac, cf_tol, & omega search, tol.
Starts to address #22
