### 2019-09-23 0f9f81b
Extend test time to 20min
