### 2019-09-23 5e89d94
Bump version number and start versioning data tarball
