### 2019-09-24 322ad38
Add options interp_only and resolve_if_found to spin sequence solver
Also add tests that these options actually do something nontrivial
