### 2019-09-24 edcde18
Put cache dir in a more intelligent place, #20
