### 2019-09-24 9368ff1
Travis caching of qnm's possible cache dirs
