### 2019-09-24 c3d56bb
More explicit checking of existence of cache location
