### 2019-09-24 8dfb3fd
Enable travis on testing branch
