+++ .travis.yml
@@ -8,7 +8,8 @@ sudo: required
 dist: xenial
 branches:
   only:
-    master
+    - master
+    - testing
 python:
   - "2.7"
   - "3.6"
