### 2019-09-24 e1fd9f6
Try using pathlib2 instead of pathlib for py2
