+++ setup.py
@@ -32,7 +32,7 @@ setuptools.setup(
         'scipy',
         'numba',
         'tqdm',
-        'pathlib ; python_version < \'3.4\'',
+        'pathlib2 ; python_version < \'3.4\'',
     ],
     classifiers=[
         "Intended Audience :: Science/Research",
