+++ qnm/schwarzschild/tabulated.py
@@ -5,7 +5,10 @@ from __future__ import division, print_function, absolute_import
 
 import logging
 import pickle
-from pathlib import Path
+try:
+    from pathlib import Path # py 3
+except ImportError:
+    from pathlib2 import Path # py 2
 
 import numpy as np
 
