### 2019-09-24 7a7efcc
Document in README that we depend on pathlib2
