### 2019-09-24 73cc0f9
Document comparison notebook in README
