### 2019-09-24 49d5479
Update JOSS submission with ref to Travis, updated bib, acks.
