+++ docs/paper/paper.bib
@@ -144,8 +144,13 @@
       author         = "Isi, Maximiliano and Giesler, Matthew and Farr, Will M.
                         and Scheel, Mark A. and Teukolsky, Saul A.",
       title          = "{Testing the no-hair theorem with GW150914}",
-      journal        = "arXiv:\href{https://arxiv.org/abs/1905.00869}{1905.00869}",
+      journal        = "Phys. Rev. Lett.",
+      volume         = "123",
       year           = "2019",
+      number         = "11",
+      pages          = "111102",
+      doi            = "10.1103/PhysRevLett.123.111102,
+                        10.1103/physrevlett.123.111102",
       eprint         = "1905.00869",
       archivePrefix  = "arXiv",
       primaryClass   = "gr-qc",
