### 2019-09-24 #1683 @duetosymmetry
Release [v0.4.0](https://github.com/duetosymmetry/qnm/releases/tag/v0.4.0):
- Rootfinding tolerance and continued fraction tolerances are now independent
- Cache directory location determined by a more standard procedure and can be controlled by env var, see docs of `qnm.cached.get_cachedir()`
- Using pathlib (or pathlib2) for cross-platform path access
- Add options `interp_only` and `resolve_if_found` to `KerrSpinSeq.__call__` for better control of (re)solving
- Start versioning cache data
- Start using Travis for automated tests
- Add notebook for comparison against Greg Cook's precomputed tables and discussion in README
