+++ docs/paper/paper.bib
@@ -149,8 +149,7 @@
       year           = "2019",
       number         = "11",
       pages          = "111102",
-      doi            = "10.1103/PhysRevLett.123.111102,
-                        10.1103/physrevlett.123.111102",
+      doi            = "10.1103/PhysRevLett.123.111102",
       eprint         = "1905.00869",
       archivePrefix  = "arXiv",
       primaryClass   = "gr-qc",
