### 2019-09-25 #1683 @xuanxu
@duncanmmacleod: I see both issues you opened in the target repo have been addressed, but there's an unchecked item in the review checklist. Can you update us on your review status?
