### 2019-09-29 75c7f0f
Make descriptions consistent in more places
