+++ docs/index.rst
@@ -1,6 +1,9 @@
 qnm's documentation: index
 ==========================
 
+`qnm` is an open source python package for computing Kerr quasinormal
+mode frequencies, separation constants, and spherical-spheroidal
+mixing coefficients.
 For a quick start, take a look at the `usage section <README.html#usage>`_ of the Welcome node.
 
 .. toctree::
