### 2019-09-30 #1683 @xuanxu
Ok, everything looks good @duetosymmetry, here are the next steps:

  - Please archive the latest release in Zenodo
  - Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also add your ORCID.

Once you do that please report here the Zenodo DOI.
