### 2019-09-30 #1683 @xuanxu
All ready for publication :tada:
Pinging @openjournals/joss-eics for final acceptance.
