### 2019-10-01 #1683 @kyleniemeyer
Hi @duetosymmetry, just a few edits necessary in the paper:
- [ ] In the first sentence of the second paragraph, remove the `[ ]` around the reference to get the an "Author et al. (2001)"-style citation.
- [ ] In the first sentence of the third paragraph, I'd suggest changing `e.g. [@Dolan:2009nk]` to `e.g., as given by @Dolan:2009nk` (or something similar)
- [ ] later in the third paragraph, change `method of Leaver [@Leaver:1985ax]` to `method of @Leaver:1985ax`
- [ ] First sentence of fourth paragraph: change to `A refinement of this method was put forth by @Cook:2014cta (see also Appendix A of @Hughes:1999bq).`

(I'm going to stop commenting on the changes in in-text citations, but please fix the rest of those.)

- [ ] In the summary, please replace the in-text links to the mentioned codes (London, Berti) with references to standard citations
- [ ] end of 3rd paragraph in Summary: add a comma after `e.g.`

Last: this isn't a required edit, but a suggestion: you have a lot of acronyms introduced in the first paragraph, and while those may be standard in your field, from a general reader's perspective they don't really save much space and instead just make reading the text a bit harder. So, you might consider removing acronyms for most terms unless a general reader would understand them. (Though I admit I'm guilty of this plenty myself.)
