### 2019-10-01 32ad592
Editorial suggestions for JOSS paper
