### 2019-10-01 #1683 @duetosymmetry
@kyleniemeyer thanks for the suggestions! I have implemented many of them. Regarding the links to London's git repo, and Berti's web page: These resources do not have some versions of record (e.g. journal article or Zenodo archive), neither has a DOI. How should I cite these resources?
