### 2019-10-01 #1683 @kyleniemeyer
@duetosymmetry thanks! for the software refs, a typical software citation should include:
- author(s)
- title + version number
- year of release associated with the version
- DOI associated with version, if possible
- URL directing to landing page

If only some of these are available, that's ok—not many have adopted archiving with DOIs yet.
