### 2019-10-01 4cbbd6d
Promote Berti and London links to biblio refs
