+++ docs/paper/paper.bib
@@ -6,12 +6,14 @@
 @misc{BertiWeb,
   author = {Emanuele Berti},
   title = {{Berti's ringdown web page}},
+  year  = {2010},
   howpublished = {\url{https://pages.jh.edu/~eberti2/ringdown/}},
 }
 
 @misc{LondonGithub,
   author = {Lionel London},
-  title = {{London's QNM codes}},
+  title = {{London's QNM codes @ 455f2d5}},
+  year  = {2017},
   howpublished = {\url{https://github.com/llondon6/kerr_public}},
 }
 
