### 2019-10-01 #1683 @duetosymmetry
@kyleniemeyer Here's the best I could accomplish. Neither of these codes have releases or version numbers. For London's, I included the short hash of the latest (2017) commit. For Berti's, I found the earliest appearance of the file on archive.org (2010), and it is identical to the presently served version.
